requirements.txt
asgiref==3.8.1
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.1
djongo==1.3.7
dnspython==2.7.0
pymongo==3.11.4
pytz==2024.2
sqlparse==0.5.2
tzdata==2024.2
